dtype inference fix

+++ generate/base.py
@@ -106,7 +106,7 @@ def generate(
     input_position = torch.tensor(T, device=device)
     finished = torch.tensor([False] * B, device=device)
     max_generation_iters = max_returned_tokens - T + 1
-    eos_id = torch.tensor(eos_id, device=device)
+    eos_id = torch.tensor(eos_id, dtype=dtype, device=device)
     for _ in range(2, max_generation_iters):
 
         token = next_token(model, input_position, token, temperature=temperature, top_k=top_k).clone()
